Move all reading logic to its own sub-module.
